+++ src/components/super/i-block/base/index.ts
@@ -144,6 +144,14 @@ export default abstract class iBlockBase extends iBlockFriends {
 	}
 
 	/**
+	 * True if all component watchers are operating in functional mode
+	 */
+	@computed()
+	get isFunctionalWatchers(): boolean {
+		return SSR || this.isFunctional;
+	}
+
+	/**
 	 * A link to the root component
 	 */
 	get r(): this['Root'] {
@@ -448,12 +456,12 @@ export default abstract class iBlockBase extends iBlockFriends {
 		optsOrHandler: AsyncWatchOptions | RawWatchHandler<this, T>,
 		handlerOrOpts?: RawWatchHandler<this, T> | AsyncWatchOptions
 	): void {
-		const
-			{async: $a} = this;
-
 		if (SSR) {
 			return;
 		}
+
+		const
+			{async: $a} = this;
 
 		let
 			handler: RawWatchHandler<this, T>,
@@ -665,7 +673,7 @@ export default abstract class iBlockBase extends iBlockFriends {
 		}
 
 		if (!this.meta.params.root) {
-			['app', 'appId', 'hydrationStore', 'ssrState'].forEach((prop) => {
+			['app', 'hydrationStore'].forEach((prop) => {
 				Object.defineProperty(this, prop, {
 					enumerable: true,
 					configurable: true,