+++ src/components/super/i-block/providers/index.ts
@@ -12,7 +12,9 @@
  */
 
 import symbolGenerator from 'core/symbol';
+
 import Provider, { providers, instanceCache, ProviderOptions } from 'core/data';
+import { unwrap as unwrapWatcher } from 'core/object/watch';
 
 import SyncPromise from 'core/promise/sync';
 import config from 'config';
@@ -298,8 +300,8 @@ export default abstract class iBlockProviders extends iBlockState {
 		opts = {
 			...opts,
 			i18n: this.i18n.bind(this),
-			id: this.r.appId,
-			remoteState: this.remoteState
+			id: this.remoteState.appProcessId,
+			remoteState: Object.cast(unwrapWatcher(this.remoteState))
 		};
 
 		let