+++ src/components/super/i-static-page/i-static-page.ts
@@ -18,24 +18,13 @@ import { Xor128 } from 'core/random/xor128';
 import { RestrictedCache } from 'core/cache';
 import { instanceCache } from 'core/data';
 
-import { setLocale, locale } from 'core/i18n';
-
 import type { AppliedRoute, InitialRoute } from 'core/router';
 import type PageMetaData from 'core/page-meta-data';
 
-import * as cookies from 'core/cookies';
 import CookieStorage from 'core/kv-storage/engines/cookie';
 
 import { SystemThemeExtractorWeb } from 'components/super/i-static-page/modules/theme';
-
-import {
-
-	resetComponents,
-
-	GlobalEnvironment,
-	ComponentResetType
-
-} from 'core/component';
+import { resetComponents, ComponentResetType } from 'core/component';
 
 import type bRouter from 'components/base/b-router/b-router';
 import type iBlock from 'components/super/i-block/i-block';
@@ -88,8 +77,8 @@ export default abstract class iStaticPage extends iPage {
 	/**
 	 * A module to work with data of data providers globally
 	 */
-	@system(() => createProviderDataStore(new RestrictedCache(10)))
-	readonly providerDataStore!: ProviderDataStore;
+	@system(() => SSR ? null : createProviderDataStore(new RestrictedCache(10)))
+	readonly providerDataStore?: ProviderDataStore;
 
 	/**
 	 * A module to manage app themes from the Design System
@@ -98,7 +87,7 @@ export default abstract class iStaticPage extends iPage {
 		o,
 		{
 			themeStorageEngine: new CookieStorage('v4ls', {
-				cookies: cookies.from(o.remoteState.cookies ?? cookies.createCookieStore('')),
+				cookies: o.remoteState.cookies,
 				maxAge: 2 ** 31 - 1
 			}),
 
@@ -134,25 +123,6 @@ export default abstract class iStaticPage extends iPage {
 	initialRoute?: InitialRoute | this['CurrentPage'];
 
 	/**
-	 * An object whose properties will extend the global object.
-	 * For example, for SSR rendering, the proper functioning of APIs such as `document.cookie` or `location` is required.
-	 * Using this object, polyfills for all necessary APIs can be passed through.
-	 */
-	@system<iStaticPage>({
-		atom: true,
-		init: (o) => o.initGlobalEnv(o.remoteState)
-	})
-
-	globalEnv!: GlobalEnvironment;
-
-	/**
-	 * A module for manipulating page metadata, such as the page title or description
-	 */
-	get pageMetaData(): PageMetaData {
-		return this.r.remoteState.pageMetaData;
-	}
-
-	/**
 	 * The name of the active route page
 	 */
 	@computed({cache: true, dependencies: ['route.meta.name']})
@@ -176,30 +146,16 @@ export default abstract class iStaticPage extends iPage {
 		this.emit('setRoute', value);
 	}
 
-	/**
-	 * The application locale
-	 */
-	get locale(): Language {
-		return this.field.get<Language>('localeStore')!;
-	}
-
-	/**
-	 * Sets a new application locale
-	 * @param value
-	 */
-	set locale(value: Language) {
-		this.field.set('localeStore', value);
-
-		try {
-			document.documentElement.setAttribute('lang', value);
-		} catch {}
-
-		setLocale(value);
-	}
-
 	override get randomGenerator(): IterableIterator<number> {
 		this[$$.randomGenerator] ??= new Xor128(19881989);
 		return this[$$.randomGenerator];
+	}
+
+	/**
+	 * A module for manipulating page metadata, such as the page title or description
+	 */
+	get pageMetaData(): PageMetaData {
+		return this.r.remoteState.pageMetaData;
 	}
 
 	/**
@@ -220,31 +176,6 @@ export default abstract class iStaticPage extends iPage {
 	 */
 	@system()
 	protected routerStore?: this['Router'];
-
-	/** {@link iStaticPage.locale} */
-	@field(() => {
-		const
-			lang = locale.value;
-
-		if (Object.isTruly(lang)) {
-			try {
-				const
-					el = document.documentElement;
-
-				if (lang != null) {
-					el.setAttribute('lang', lang);
-
-				} else {
-					el.removeAttribute('lang');
-				}
-
-			} catch {}
-		}
-
-		return lang;
-	})
-
-	protected localeStore!: string;
 
 	/**
 	 * Cache of root modifiers
@@ -404,34 +335,18 @@ export default abstract class iStaticPage extends iPage {
 		this.shouldMountTeleports = true;
 	}
 
-	/**
-	 * Synchronization of the `localeStore` field
-	 * @param locale
-	 */
-	@watch(['localeStore', 'globalEmitter:i18n.setLocale'])
-	protected syncLocaleWatcher(locale: Language): void {
-		if (this.locale === locale) {
-			return;
-		}
-
-		this.locale = locale;
-		this.forceUpdate().catch(stderr);
-	}
-
 	protected override beforeDestroy(): void {
 		super.beforeDestroy();
 		this.hydrationStore?.clear();
 
-		Object.forEach(this.ssrState, (_, key: string, state) => {
-			delete state![key];
-		});
-
 		const
-			isThisApp = new RegExp(RegExp.escape(`:${RegExp.escape(this.appId)}:`));
+			isThisApp = new RegExp(RegExp.escape(`:${RegExp.escape(this.remoteState.appProcessId)}:`));
 
 		Object.forEach(instanceCache, (provider, key) => {
 			if (isThisApp.test(key)) {
-				provider.destroy();
+				// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+				provider?.destroy();
+				delete instanceCache[key];
 			}
 		});
 	}