+++ src/core/component/reflect/constructor.ts
@@ -65,7 +65,8 @@ export function getInfoFromConstructor(
 	declParams?: ComponentOptions
 ): ComponentConstructorInfo {
 	const
-		name = declParams?.name ?? getComponentName(constructor);
+		name = declParams?.name ?? getComponentName(constructor),
+		layer = declParams?.layer;
 
 	const
 		parent = Object.getPrototypeOf(constructor),
@@ -115,6 +116,8 @@ export function getInfoFromConstructor(
 
 	return {
 		name,
+		layer,
+
 		componentName: name.replace(isSmartComponent, ''),
 		constructor,
 		params,