+++ src/core/index.ts
@@ -10,10 +10,15 @@ import '@v4fire/core/core';
 import { resolveAfterDOMLoaded } from 'core/event';
 
 import initApp from 'core/init';
-import createInitAppSemaphore from 'core/init/semaphore';
+
+import * as session from 'core/session';
+import SessionEngine from 'core/session/engines';
 
 export * as cookies from 'core/cookies';
-export { initApp, createInitAppSemaphore };
+export * as session from 'core/session';
+export * as kvStorage from 'core/kv-storage';
+
+export { initApp };
 
 //#unless runtime has storybook
 
@@ -25,10 +30,48 @@ if (SSR) {
 		.then(() => {
 			const
 				targetToMount = document.querySelector<HTMLElement>('[data-root-component]'),
-				rootComponentName = targetToMount?.getAttribute('data-root-component'),
-				ready = createInitAppSemaphore();
+				rootComponentName = targetToMount?.getAttribute('data-root-component');
 
-			return initApp(rootComponentName, {targetToMount, ready});
+			return initApp(rootComponentName, {
+				appProcessId: Object.fastHash(Math.random()),
+
+				cookies: document,
+				session: session.from(SessionEngine),
+				location: getLocationAPI(),
+
+				targetToMount
+			});
+
+			function getLocationAPI(): URL {
+				Object.defineProperties(location, {
+					username: {
+						configurable: true,
+						enumerable: true,
+						get: () => ''
+					},
+
+					password: {
+						configurable: true,
+						enumerable: true,
+						get: () => ''
+					},
+
+					searchParams: {
+						configurable: true,
+						enumerable: true,
+						get: () => new URLSearchParams(location.search)
+					},
+
+					toJSON: {
+						configurable: true,
+						enumerable: true,
+						writable: false,
+						value: () => location.toString()
+					}
+				});
+
+				return Object.cast(location);
+			}
 		})
 
 		.catch(stderr);